+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Links to the unofficial discord, the Workshop and the subreddit
+- Option for automatic install of D-U and DMC-core
+
 ### Changed
 - Improved error messages
+- Code improvements
 
 ## [2.0.0] - 2017-08-18
 ### Added